+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.13.0] - (15.04.2025)
+
+## Changes
+
+- added pdm energy and production data to RESTFul and MQTT for Envoy FW >= 8.2.4xx
+- removed production all from RESTFul and MQTT
+- redme update
+- cleanup
+
 ## [9.12.6] - (08.04.2025)
 
 ## Changes